Add a `namespace` class attribute to ParameterSet.
Also adds a mechanism for convenience methods for common parsing 
contexts which extend `namespace`, and uses this to define a 'numpy' 
context.
    
The `namespace` attribute is used as the default namespace in `__init__` 
for parsing strings.
By storing it as a class attribute, rather than inlined within
`read_from_str`, it is modifiable by user code. An example use case is
mapping `array` to `numpy.array` to allow parsing of numpy arrays.
Although `read_from_str` already provides the `update_namespace`
argument, libraries which call `ParameterSet` typically do so without
setting this argument, leaving no way for the user to update the
namespace. This happens for example within Sumatra's `new_record` 
method.
And in any case, specifying the namespace every time we parse a
ParameterSet is much more error prone than setting the default once.