Less confusing error message for incorrect path to local parameter file
Parameter files are first attempted to be opened locally, and if that 
fails, then attempted to be interpreted as a URL.
This means that, when one mistypes the location of a local parameter 
file, the result is an error from urllib complaining about an unknown 
url type. The reason (and proper resolution) for this error is virtually 
impossible to understand without looking at the code.
This commit catches the error by urllib, and reraises with a more useful 
message that contains both possible reasons for failure (wrong path, 
incorrect URL)